Refactor PromptResponse into types folder
