+++ js/packages/teams-ai/src/models/PromptCompletionModel.ts
@@ -9,6 +9,7 @@
 import { TurnContext } from 'botbuilder-core';
 import { Message, PromptFunctions, PromptTemplate } from '../prompts';
 import { Tokenizer } from '../tokenizers';
+import { PromptResponse } from '../types';
 import { Memory } from '../MemoryFork';
 import StrictEventEmitter from '../external/strict-event-emitter-types';
 import EventEmitter from 'events';
@@ -35,7 +36,7 @@ export interface PromptCompletionModelEvents {
      * @param chunk Message delta received from the model.
      */
     chunkReceived: (context: TurnContext, memory: Memory, chunk: PromptChunk) => void;
-    
+
     /**
      * Triggered after the model finishes returning a response.
      * @param context Current turn context.
@@ -90,45 +91,6 @@ export interface PromptCompletionModel {
         tokenizer: Tokenizer,
         template: PromptTemplate
     ): Promise<PromptResponse<string>>;
-}
-
-/**
- * Status of the prompt response.
- * @remarks
- * `success` - The prompt was successfully completed.
- * `error` - An error occurred while completing the prompt.
- * `rate_limited` - The request was rate limited.
- * `invalid_response` - The response was invalid.
- * `too_long` - The rendered prompt exceeded the `max_input_tokens` limit.
- */
-export type PromptResponseStatus = 'success' | 'error' | 'rate_limited' | 'invalid_response' | 'too_long';
-
-/**
- * Response returned by a `PromptCompletionClient`.
- * @template TContent Optional. Type of the content in the message. Defaults to `unknown`.
- */
-export interface PromptResponse<TContent = unknown> {
-    /**
-     * Status of the prompt response.
-     */
-    status: PromptResponseStatus;
-
-    /**
-     * User input message sent to the model. `undefined` if no input was sent.
-     */
-    input?: Message<any>;
-
-    /**
-     * Message returned.
-     * @remarks
-     * This will be a `Message<TContent>` object if the status is `success`, otherwise it will be a `string`.
-     */
-    message?: Message<TContent>;
-
-    /**
-     * Error returned.
-     */
-    error?: Error;
 }
 
 /**
